api/requirements.txt
tornado==4.3
psycopg2==2.6.2
pylint==1.5.6
momoko==2.2.3
yoyo-migrations==5.0.3
